// packages/talmud/src/client/shapes.ts
export interface Rabbi {
  name: string;
  nameHe: string;
  period: string;
  location: string;
  role: string;
  opinionStart?: string;
}

export interface Section {
  title: string;
  summary: string;
  excerpt?: string;
  rabbis: Rabbi[];
  startSegIdx?: number;
  endSegIdx?: number;
}